Move RemoteAuthentication and AuthenticationTypes out of WinPcap and Npcap and into LibPcap.
Note this is an API breaking change
